Board set-up outline and test outline.
This commit does not include fully functional code. It simply outlines
the direction for future code.

The board is going to hold a lot of the state of the world (with the
exception of player hands and pieces). This board is going to be
represented by a series of nodes in a networkx graph. These nodes will
be of different types, outlined in this commit.

This commit also contains a fair amount of working notes, and working
method names.

Future commits should build on this structure.