Simplified board_types to use standard hex_utils.
Since the `StandardBoardType` is a straightforward mapping of ordinals
to hexagons and back, similar to those refactored into `hex_utils`,
the code used to compute this can be offloaded almost entirely to the
`hex_utils` module.

The caching is left in place here, rather than in `hex_utils` so that
`hex_utils` can stay pure. As a library, it should be further
optimized.

This is the final commit in this branch.