Move robber and import restructuring.
The main functional change of this commit is the
`board.Board.move_robber` method. However, in attempting to test this
new method, I uncovered problems with how modules were being imported,
and decided to move to a system importing as if the package is already
loaded from the entry point.

Additionally I added a __repr__ method to board.Tile.